@@ -1,47 +1,49 @@
-apply plugin: 'com.android.application'
-//apply plugin: 'realm-android'
+plugins {
+    id 'com.android.application'
+    id 'com.google.gms.google-services'
+    id 'com.google.firebase.crashlytics'
+}
 
 android {
-    compileSdkVersion 26
-    buildToolsVersion "25.0.3"
+    compileSdk 31
+
     defaultConfig {
         applicationId "com.ak.pesgm"
-        minSdkVersion 14
-        targetSdkVersion 26
+        minSdk 21
+        targetSdk 31
         versionCode 5
         versionName "3.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
-        vectorDrawables.useSupportLibrary = true
+
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
+
     buildTypes {
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
+    }
 }
 
 dependencies {
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
-        exclude group: 'com.android.support', module: 'support-annotations'
-    })
-
-
-    implementation 'com.android.support:appcompat-v7:26.+'
-    implementation 'com.android.support:design:26.+'
-    implementation 'com.android.support:support-vector-drawable:26.+'
-    implementation 'com.jakewharton:butterknife:8.8.0'
-    implementation 'com.android.support:support-v4:26.+'
-    implementation 'com.android.support:cardview-v7:25.0.1'
-    implementation 'com.github.bumptech.glide:glide:3.7.0'
-
-    // compile 'com.jackpocket:scratchoff:1.0.2'
-
-    //  compile 'com.netflix.rxjava:rxjava-android:0.16.1'
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
+    implementation 'androidx.appcompat:appcompat:1.4.2'
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'com.google.android.material:material:1.6.1'
+    implementation 'com.google.firebase:firebase-crashlytics:18.2.11'
+    implementation 'com.google.firebase:firebase-analytics:21.1.0'
+    implementation 'com.google.firebase:firebase-messaging:23.0.6'
+    implementation 'com.jakewharton:butterknife:10.0.0'
+    implementation 'com.github.bumptech.glide:glide:4.12.0'
     implementation 'com.joanzapata.pdfview:android-pdfview:1.0.4@aar'
 
-    testCompile 'junit:junit:4.12'
-    annotationProcessor 'com.jakewharton:butterknife-compiler:8.8.0'
+    annotationProcessor 'com.jakewharton:butterknife-compiler:10.0.0'
+    annotationProcessor 'com.github.bumptech.glide:compiler:4.12.0'
+    testImplementation 'junit:junit:4.13.2'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.3'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
 }
