@@ -2,7 +2,6 @@ package com.ak.pesgm.fragment;
 
 
 import android.app.Fragment;
-import android.content.ActivityNotFoundException;
 import android.content.Context;
 import android.content.Intent;
 import android.content.pm.ApplicationInfo;
@@ -10,17 +9,8 @@ import android.content.pm.PackageManager;
 import android.content.res.AssetManager;
 import android.content.res.Configuration;
 import android.content.res.Resources;
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
-import android.graphics.Canvas;
-import android.graphics.Paint;
 import android.net.Uri;
 import android.os.Bundle;
-import android.os.Environment;
-import android.provider.MediaStore;
-import android.support.annotation.IdRes;
-import android.support.v7.widget.CardView;
-import android.text.Html;
 import android.util.DisplayMetrics;
 import android.util.Log;
 import android.view.LayoutInflater;
@@ -28,30 +18,25 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.webkit.WebSettings;
 import android.webkit.WebView;
-import android.widget.ImageView;
 import android.widget.RadioButton;
 import android.widget.RadioGroup;
-import android.widget.TextView;
+
+import androidx.annotation.IdRes;
+import androidx.cardview.widget.CardView;
 
 import com.ak.pesgm.R;
-import com.ak.pesgm.activity.MainActivity;
 import com.ak.pesgm.activity.PdfActivity;
 import com.ak.pesgm.activity.PesgmActivity;
 import com.ak.pesgm.app.SessionManager;
-//import com.jackpocket.scratchoff.ScratchoffController;
 
 import java.io.File;
-import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
-import java.util.Date;
 import java.util.Locale;
-import java.util.concurrent.TimeUnit;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
-
 
 /**
  * A simple {@link Fragment} subclass.
@@ -77,12 +62,7 @@ public class AboutFragment extends Fragment {
 
     @BindView(R.id.wvAddress)
     WebView wvAddress;
-    /* @BindView(R.id.cv_create_frame)
-     CardView cvFrame;*/
-    /*@BindView(R.id.ivPreview)
-    ImageView imageView;*/
-    /*@BindView(R.id.tvTips)
-    TextView tvTips;*/
+
 
     public static int count = 0;
     String[] tips;
@@ -92,11 +72,8 @@ public class AboutFragment extends Fragment {
     String fburl = "https://www.facebook.com/PESGMandal1/";
     String instaurl = "https://www.instagram.com/pesgm/";
 
-/*    ScratchoffController controller;*/
-    static final int CAM_REQUEST = 1;
 
     public AboutFragment() {
-        // Required empty public constructor
     }
 
 
@@ -114,10 +91,7 @@ public class AboutFragment extends Fragment {
             rbEnglish.setChecked(true);
         }
 
-
         tips = getResources().getStringArray(R.array.tips_tricks);
-
-      //  setTips();
         rgLang.setOnCheckedChangeListener(new RadioGroup.OnCheckedChangeListener() {
             @Override
             public void onCheckedChanged(RadioGroup group, @IdRes int checkedId) {
@@ -135,41 +109,10 @@ public class AboutFragment extends Fragment {
             }
         });
 
-       /* cvFrame.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Intent camera_intent = new Intent(MediaStore.ACTION_IMAGE_CAPTURE);
-                File file = getFile();
-                camera_intent.putExtra(MediaStore.EXTRA_OUTPUT, Uri.fromFile(file));
-                startActivityForResult(camera_intent,CAM_REQUEST);
-            }
-        });*/
-
         cvAartiCollection.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-
-                //CopyReadAssets();
-
-
-              /*  try {
-
-                    // your intent here
-                    Uri path = Uri.parse("android.resource://com.ak.pesgm/raw/aarti.pdf");
-                    Intent intent = new Intent(Intent.ACTION_VIEW);
-                    intent.setDataAndType(path, "application/pdf");
-                    // intent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
-                    getActivity().startActivity(intent);
-
-                } catch (ActivityNotFoundException e) {
-                    // show message to user
-
-                    Log.v("asf",""+e.getMessage());*/
-
                 startActivity(new Intent(getActivity(), PdfActivity.class));
-                //}
-
-
             }
         });
 
@@ -188,17 +131,6 @@ public class AboutFragment extends Fragment {
             }
         });
 
-
-       /* controller = new ScratchoffController(getActivity())
-                .setThresholdPercent(0.50d)
-                .setTouchRadiusDip(getActivity(), 20)
-                .setFadeOnClear(true)
-                .setClearOnThresholdReached(true)
-                .attach(view.findViewById(R.id.scratch_view), view.findViewById(R.id.scratch_view_behind));
-*/
-        //controller.setThresholdPercent(0.40d);
-        //controller.setFadeOnClear(true);
-
         wvAddress.getSettings().setJavaScriptEnabled(true);
         wvAddress.loadData("<iframe src=\"https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d943.3104563228565!2d72.81203811468214!3d18.964924621029503!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x0%3A0x632786d9ba511713!2sSachinam+Heights!5e0!3m2!1sen!2sin!4v1503522801813\" width=\"400\" height=\"300\" frameborder=\"0\" style=\"border:0\" allowfullscreen></iframe>", "text/html", "utf-8");
         WebSettings webViewSettings = wvAddress.getSettings();
@@ -209,21 +141,6 @@ public class AboutFragment extends Fragment {
 
         return view;
     }
-
-  /*  Date curentTime;
-    private File getFile()
-    {
-        curentTime = new Date();
-        String root = Environment.getExternalStorageDirectory().toString();
-        File folder = new File(root + "/pesgm");
-        if (!folder.mkdirs()) {
-            Log.e("asdf", "Directory not created");
-        }
-
-        File image_file = new File(folder,"selfie_cam"+curentTime+".jpg");
-        return image_file;
-    }*/
-
 
     public void setLocale(String lang) {
 
@@ -238,65 +155,6 @@ public class AboutFragment extends Fragment {
         refresh.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
         startActivity(refresh);
     }
-
-
-   /* @Override
-    public void onActivityResult(int requestCode, int resultCode, Intent     data) {
-        setAndSaveImageWithOverlay(getBitmapOfSnappedImage());
-    }*/
-
-
-
-  /*  public Bitmap getBitmapOfSnappedImage(){
-
-
-        String root = Environment.getExternalStorageDirectory().toString();
-        String path = root + "/pesgm"+"/selfie_cam"+curentTime+".jpg";
-
-        File image = new File(path);
-        BitmapFactory.Options bmOptions = new BitmapFactory.Options();
-        Bitmap bitmap =     BitmapFactory.decodeFile(image.getAbsolutePath(),bmOptions);
-        return bitmap;
-    }
-
-    public void setAndSaveImageWithOverlay(Bitmap snappedImage){
-        Bitmap b = Bitmap.createBitmap(snappedImage.getWidth(),     snappedImage.getHeight(), Bitmap.Config.ARGB_8888);
-
-        //the overlay png file from drawable folder
-        Bitmap overlay = BitmapFactory.decodeResource(getResources(),     R.drawable.cup);
-        overlay =     Bitmap.createScaledBitmap(overlay,snappedImage.getWidth(),snappedImage.getHeight    (),false);
-
-        //create canvas with a clean bitmap
-        Canvas canvas = new Canvas(b);
-        //draw the snappedImage on the canvas
-        canvas.drawBitmap(snappedImage, 0, 0, new Paint());
-        //draw the overlay on the canvas
-        canvas.drawBitmap(overlay, 0, 0, new Paint());
-
-        imageView.setImageBitmap(b);
-
-        SaveImage(b);
-    }
-
-
-    private void SaveImage(Bitmap finalBitmap) {
-
-        String root = Environment.getExternalStorageDirectory().toString();
-        File myDir = new File(root + "/Blue_&_Gold_Perade");
-        myDir.mkdirs();
-        String fname = "selfie_cam"+curentTime+".jpg";
-        File file = new File (myDir, fname);
-        if (file.exists ()) file.delete ();
-        try {
-            FileOutputStream out = new FileOutputStream(file);
-            finalBitmap.compress(Bitmap.CompressFormat.JPEG, 90, out);
-            out.flush();
-            out.close();
-
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-    }*/
 
     public Intent getFacebookIntent(String url) {
 
@@ -358,51 +216,5 @@ public class AboutFragment extends Fragment {
             out.write(buffer, 0, read);
         }
     }
-
-
-   /* public void setTips() {
-        long lastLoginTime = sessionManager.getLastTipsTime();
-        Date dt = new Date(lastLoginTime);
-        long diff = getDateDiff(dt, new Date(), TimeUnit.HOURS);
-
-        if (diff >= 8) {
-
-            if (count >= tips.length) {
-                count = 0;
-            } else {
-                count = sessionManager.getTipsCount() + 1;
-            }
-            sessionManager.setTipsTimeAndCount(count);
-
-            tvTips.setText(tips[sessionManager.getTipsCount()]);
-
-        } else {
-            tvTips.setText(tips[sessionManager.getTipsCount()]);
-        }
-    }
-
-    public long getDateDiff(Date date1, Date date2, TimeUnit timeUnit) {
-        long diffInMillies = date2.getTime() - date1.getTime();
-        Log.v("Date Difference", "is : " + timeUnit.convert(diffInMillies, TimeUnit.MILLISECONDS));
-        return timeUnit.convert(diffInMillies, TimeUnit.MILLISECONDS);
-    }
-
-    @Override
-    public void onPause() {
-        controller.onPause();
-        super.onPause();
-    }
-
-    @Override
-    public void onResume() {
-        super.onResume();
-        controller.onResume();
-    }
-
-    @Override
-    public void onDestroy() {
-        controller.onDestroy();
-        super.onDestroy();
-    }*/
 
 }
